Flower on tree dies when damaging tree

+++ Assets/Scripts/UnitsSharedScripts/Health.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class Health : MonoBehaviour
 {
-    public float currentHealth;
-    public float maxHealth;
+    public int currentHealth;
+    public int maxHealth;
     
     // Start is called before the first frame update
     void Start()
